+++ test/unit/test_session.py
@@ -1182,3 +1182,21 @@ def test_context_manager(session):
         assert session.closed is False
 
     assert session.closed is True
+
+
+def test_delayed_job(session):
+    '''Test the delayed_job action'''
+
+    result = session.delayed_job(
+        ftrack_api.symbol.JOB_SYNC_USERS_LDAP
+    )
+
+    assert isinstance(
+        result, ftrack_api.entity.job.Job
+    )
+
+    with pytest.raises(ValueError):
+        session.delayed_job(
+            'DUMMY_JOB'
+        )
+