+++ doc/release/release_notes.rst
@@ -19,7 +19,8 @@ Release Notes
     .. change:: new
         :tags: task template
 
-        Added example for managing task templates through the API. See :ref:`example/task_template` for example usage.
+        Added example for managing task templates through the API. See
+        :ref:`example/task_template` for example usage.
 
 .. release:: 1.3.1