+++ doc/release/release_notes.rst
@@ -9,11 +9,18 @@ Release Notes
 
 .. currentmodule:: ftrack_api.session
 
-.. release:: Upcoming
+.. release:: upcoming
+
+    .. change:: fixed
+        :tags: metadata
+
+        Not possible to set metadata on creation.
+
     .. change:: new
         :tags: task template
 
-        Added example for managing task templates through the API. See :ref:`example/task_template` for example usage.
+        Added example for managing task templates through the API. See
+        :ref:`example/task_template` for example usage.
 
 .. release:: 1.3.1